+++ set.go
@@ -327,7 +327,7 @@ func (s Set) Powerset() Set {
 	elems := s.Elements()
 	subset := Set{}
 	result := NewSet(subset)
-	for i := BitIterator(1); i < 1<<n; i++ {
+	for i := BitIterator(1); i < 1<<uint(n); i++ {
 		// Use a special counting order that flips a single bit at at time. The
 		// bit to flip is the same as the lowest-order 1-bit in the normal
 		// counting order, denoted by `(1)`. The flipped bit's new value is the
